Fix Encapsulation Problem
In file: ClientObject.java, class:  has a field that is declared as public but it may allow unwarranted access. The access to the field should be restricted and should only be through accessor methods. iCR suggested changes in 6 files to resolve the problem. For more information, read CERT Secure coding standard OBJ01-J